pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "3DPPE"
version = "2.0.0"
description = "Modern 3D Point Positional Encoding for Multi-Camera 3D Object Detection"
authors = [
    {name = "Changyong Shu", email = "changyong.shu@example.com"},
    {name = "Jiajun Deng", email = "jiajun.deng@example.com"},
]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.9"
keywords = ["computer-vision", "3d-detection", "transformers", "pytorch"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
    "torchmetrics>=1.0.0",
    "lightning>=2.0.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "wandb>=0.15.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "tqdm>=4.65.0",
    "rich>=13.0.0",
    "einops>=0.6.0",
    "timm>=0.9.0",
    "nuscenes-devkit>=1.1.0",
    "lyft-dataset-sdk>=0.0.8",
    "networkx>=3.0",
    "plyfile>=1.0.0",
    "trimesh>=4.0.0",
    "numba>=0.57.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.5.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "myst-parser>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/your-repo/3DPPE"
Documentation = "https://3DPPE.readthedocs.io/"
Repository = "https://github.com/your-repo/3DPPE"
Issues = "https://github.com/your-repo/3DPPE/issues"
Changelog = "https://github.com/your-repo/3DPPE/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "nuscenes.*",
    "lyft_dataset_sdk.*",
    "mmcv.*",
    "mmdet.*",
    "mmdet3d.*",
]
ignore_missing_imports = true
